Added new features

+++ app/src/main/java/com/tahayasindogukan/studentclubapplication/core/entitiy/Request.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 data class Request(
     val name: String="",
     val surname: String="",
-    val visibility: Int=0
+    val visibility: Long = 0
 ) : Serializable {
 }
